requirements.txt
boto3==1.5.14
botocore==1.8.28
Django==1.8.5
django-ckeditor==5.4.0
django-disqus==0.5
django-js-asset==0.1.1
django-storages==1.6.5
docutils==0.14
funcsigs==1.0.2
futures==3.2.0
jmespath==0.9.3
mock==2.0.0
mysql-python==1.2.3
pbr==3.1.1
Pillow==2.8.2
python-dateutil==2.6.1
pytz==2017.2
s3transfer==0.1.12
six==1.11.0